pyproject.toml
[tool.poetry]
name = "brainiax"
version = "0.1"
description = "brainiax : Your Personalised Academic Aid AI chatbot"
authors = ["Zuhair"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = { extras = ["all"], version = "^0.110.0" }
python-multipart = "^0.0.9"
injector = "^0.21.0"
pyyaml = "^6.0.1"
transformers = "^4.38.2"
llama-index-core = "^0.10.14"
llama-index-readers-file = "^0.1.6"
llama-index-llms-ollama = "^0.1.2"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-vector-stores-qdrant = "^0.1.3"
gradio = "^4.19.2"

